Some fixes in chapter 2.
	Add: missing excercise.
	Fix: change distance to double in 2.
	Fix: missing include in 4.
	Fix: change convert to double in 6.

+++ chapter2/2.cpp
@@ -8,7 +8,7 @@
 int main()
 {
 	using namespace std;
-	int dist = 0;
+	double dist = 0;
 
 	cout << "Please, enter distance in furlongs:";
 	cin >> dist;
